complete fix

+++ ffn/swin_transformer_ln_qkv_quant/op_host/swin_transformer_ln_qkv_quant_tiling.cpp
@@ -265,13 +265,13 @@ ge::graphStatus SwinTransformerLnQkvQuantTilingCompute::GetBaseParams(gert::Tili
                     "The attrs param is null!"),
                     return ge::GRAPH_FAILED);
     size_t idx = 0;
-    headNum = *attrs->GetAttrPointer<int>(idx++);
-    auto seqLength = attrs->GetAttrPointer<int>(idx++);
+    headNum = *attrs->GetAttrPointer<int64_t>(idx++);
+    auto seqLength = attrs->GetAttrPointer<int64_t>(idx++);
     auto epsilonOpt = attrs->GetAttrPointer<float>(idx++);
-    oriHeight = *attrs->GetAttrPointer<int>(idx++);
-    oriWeight = *attrs->GetAttrPointer<int>(idx++);
-    hWinSize = *attrs->GetAttrPointer<int>(idx++);
-    wWinSize = *attrs->GetAttrPointer<int>(idx++);
+    oriHeight = *attrs->GetAttrPointer<int64_t>(idx++);
+    oriWeight = *attrs->GetAttrPointer<int64_t>(idx++);
+    hWinSize = *attrs->GetAttrPointer<int64_t>(idx++);
+    wWinSize = *attrs->GetAttrPointer<int64_t>(idx++);
     bTransOpt = *attrs->GetAttrPointer<bool>(idx++);
     sizePerHead = (seqLength != nullptr) ? (*seqLength) : sizePerHead;
     transposeB = bTransOpt;
